app/pyproject.toml
[project]
name = "crossy"
version = "1.0.0"
description = "Border Patrol Interior Checkpoint Crossing Simulator"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "motor>=3.3.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
]
